LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing-deprecated.html

<!DOCTYPE html>
<html>
<head>
<script src='resources/multiple-iframe-test.js'></script>
<script>
var tests = [
    ['yes', 'allow \'none\'', 'resources/script.js'],
    ['yes', 'options \'inline-script\'', 'resources/script.js'],
    ['yes',  'policy-uri http://example.com/policy', 'resources/script.js'],
];
</script>
</head>
<body onload="test()">
<p>
  Test that directives that predate the CSP 1.0 standard generate proper
  warnings to set developer expectations correctly.
</p>
