Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing-deprecated-expected.txt

CONSOLE MESSAGE: The 'allow' directive has been replaced with 'default-src'. Please use that directive instead, as 'allow' has no effect.
CONSOLE MESSAGE: The 'options' directive has been replaced with 'unsafe-inline' and 'unsafe-eval' source expressions for the 'script-src' and 'style-src' directives. Please use those directives instead, as 'options' has no effect.
CONSOLE MESSAGE: The 'policy-uri' directive has been removed from the specification. Please specify a complete policy via the Content-Security-Policy header.
Test that directives that predate the CSP 1.0 standard generate proper warnings to set developer expectations correctly.



--------
Frame: '<!--frame1-->'
--------
PASS

--------
Frame: '<!--frame2-->'
--------
PASS

--------
Frame: '<!--frame3-->'
--------
PASS
